requirements.txt
tensorflow-gpu==2.12.0  # Or the latest compatible version
Flask==2.2.2
scapy==2.4.5
numpy==1.23.5
psutil==5.9.4
GPUtil==1.4.0
tqdm==4.64.1